feat: add findArgs to utilities
